% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distances.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Get the haversine distances and elevation range from coordinates and elevation}
\usage{
calculate_distances(
  df,
  coord_names = c("long", "lat"),
  elevation_name = "altitud"
)
}
\arguments{
\item{df}{A data frame}

\item{coord_names}{column names from the coordinates like: c(long, lat)}

\item{elevation_name}{column name for the elevation}
}
\description{
Get the haversine distances and elevation range from coordinates and elevation
}
\examples{
data(Huancavelica_2013)
}
