% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocf.R
\name{CCF}
\alias{CCF}
\title{Calculate cumulative cultivar frequency}
\usage{
CCF(dfr, vname, hh, community, location, pctn = TRUE)
}
\arguments{
\item{dfr}{A data frame}

\item{vname}{variety names}

\item{hh}{household names or codes}

\item{community}{community}

\item{location}{location}

\item{pctn}{Boolean whether calculate percentages}
}
\value{
A data frame with the CCf values
}
\description{
Calculate cumulative cultivar frequency
}
\examples{
data(varieties_data)
ccf_data <- CCF(dfr=varieties_data,
vname="variety_name",
hh="household_code",
community="community",
location="location")
}
