% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Huancavelica_2013}
\alias{Huancavelica_2013}
\title{Example dataset for OCF and RCF calculation Huancavelica 2013}
\format{
\subsection{'varieties_data'}{

A data frame with columns:
\describe{
\item{variety_name}{Name of the variety}
\item{household_code}{Household identifier}
\item{number_of_tubers}{Number of varieties sampled per household}
\item{community}{Community identifier}
\item{location}{Location identifier}
}
}
}
\usage{
data(Huancavelica_2013)
}
\description{
Example dataset for OCF and RCF calculation Huancavelica 2013
}
\keyword{datasets}
