% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocf.R
\name{OCF}
\alias{OCF}
\title{Calculate overall cultivar frequency (OCF)}
\usage{
OCF(dfr, vname, hh, community, location, quantiles = c(0.25, 0.5, 0.75))
}
\arguments{
\item{dfr}{A data frame}

\item{vname}{variety names}

\item{hh}{household names or codes}

\item{community}{community names}

\item{location}{location of the varieties}

\item{quantiles}{quantiles to get}
}
\value{
A data frame with OCF percentages and metrics to calculate the OCF
}
\description{
Calculate overall cultivar frequency (OCF)
}
\examples{
#' data(varieties_data)
ocf_data <- OCF(dfr=varieties_data,
vname="variety_name",
hh="household_code",
community="community",
location="location")
}
