% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Red_Listing.R
\name{Plot_Variable_Red_Listing}
\alias{Plot_Variable_Red_Listing}
\title{Plot the RCF vs OCF values to identify Red Listing varieties}
\usage{
Plot_Variable_Red_Listing(
  OCF_df,
  RCF_df,
  quantiles = c(0.25, 0.5, 0.75),
  type = "log"
)
}
\arguments{
\item{OCF_df}{A data frame generated by the OCF function}

\item{RCF_df}{A data frame generated by the RCF function}

\item{quantiles}{Which quantiles you want to break the RCf and OCF distributions}

\item{type}{Normal scale or Logarithm scale. Default value is log scale.}
}
\value{
A plot where we can see the RCF vs OCF in normal or log scale.
}
\description{
Plot the RCF vs OCF values to identify Red Listing varieties
}
\examples{
data(varieties_data)
ocf_data <- OCF(dfr=varieties_data,
vname="variety_name",
hh="household_code",
community="community",
location="location")
rcf_data <- RCF(dfr=varieties_data,
vname="variety_name",
hh="household_code",
nsvarie="number_of_tubers",
community="community",
location="location")
Plot_Variable_Red_Listing(ocf_data, rcf_data)
}
