% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf.R
\name{RCF}
\alias{RCF}
\title{Calculate relative cultivar frequency (RCF)}
\usage{
RCF(dfr, vname, hh, nsvarie, community, location, is_grouped = TRUE)
}
\arguments{
\item{dfr}{A data frame}

\item{vname}{variety names}

\item{hh}{household names or codes}

\item{nsvarie}{number of varieties sampled in the field per household}

\item{community}{community}

\item{location}{location}

\item{is_grouped}{Whether we calculate the RCF by location, default=TRUE}
}
\value{
A data frame with RCF percentages and metrics to calculate the RCF
}
\description{
Calculate relative cultivar frequency (RCF)
}
\examples{
data(varieties_data)
rcf_data <- RCF(dfr=varieties_data,
vname="variety_name",
hh="household_code",
nsvarie="number_of_tubers",
community="community",
location="location")
}
