% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocf.R
\name{CCF}
\alias{CCF}
\title{Calculate cumulative cutlivar frequency}
\usage{
CCF(dfr, vname, hh, community, location)
}
\arguments{
\item{dfr}{A data frame}

\item{vname}{variety names}

\item{hh}{household names or codes}

\item{community}{community}

\item{location}{location}
}
\value{
A data frame with the CCf values
}
\description{
Calculate cumulative cutlivar frequency
}
