% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndangeRed-package.R
\docType{package}
\name{EndangeRed-package}
\alias{EndangeRed}
\alias{EndangeRed-package}
\title{EndangeRed: Identify Varieties in Danger of Loss}
\description{
Functions to calculate overall cultivar frequency, relative cultivar frequency and plot the relation between these two.
}
\author{
\strong{Maintainer}: Piero Palacios \email{piero.palacios@cgiar.org}

Authors:
\itemize{
  \item Henry Juarez \email{h.juarez@cgiar.org}
  \item Stef De Haan
}

}
\keyword{internal}
