% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcf.R
\name{HCF}
\alias{HCF}
\title{Calculate household cultivar frequency}
\usage{
HCF(dfr, hh, nsvarie, pctn = TRUE)
}
\arguments{
\item{dfr}{A data frame}

\item{hh}{household names or codes}

\item{nsvarie}{number of varieties sampled in the field per household}

\item{pctn}{If we calculate the RCF as percentage, default=TRUE}
}
\value{
A data frame with the HCf values
}
\description{
Calculate household cultivar frequency
}
\examples{
data(Huancavelica_2013)
hcf_data <- HCF(dfr=Huancavelica_2013,
hh="household",
nsvarie="cantidad")
}
