% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocf.R
\name{OCF}
\alias{OCF}
\title{Calculate overall cultivar frequency (OCF)}
\usage{
OCF(dfr, vname, hh, community, location)
}
\arguments{
\item{dfr}{A data frame}

\item{vname}{variety names}

\item{hh}{household names or codes}

\item{community}{community names}

\item{location}{location of the varieties}
}
\value{
A data frame with OCF percentages and metrics to calculate the OCF
}
\description{
Calculate overall cultivar frequency (OCF)
}
\examples{
data(Huancavelica_2013)
ocf_data <- OCF(dfr=Huancavelica_2013,
vname="final_variety_name",
hh="household",
community="comunidad",
location="region")
}
